+++ src/app/page.tsx
@@ -1,12 +1,9 @@
 import { PostCard } from "@/components";
+import { prisma } from "@/lib";
 import { type DummyJSONPost } from "@/types";
 
 const Page = async (): Promise<JSX.Element> => {
-  const response: Response = await fetch(
-    "https://dummyjson.com/posts?limit=0&select=id,body,title"
-  );
-  const data: { posts: DummyJSONPost[] } = await response.json();
-  const posts: DummyJSONPost[] = data.posts;
+  const posts: DummyJSONPost[] = await prisma.post.findMany();
 
   return (
     <main>
